Demonstrate it can be linked to
The app that links must
1. Load from server whenever it gets /appointment/somehost/start
2. The server must serve the appointment-app on such a request. It's the
   servers responsibility to authenticate. The server must provide its
name as a data-attribute. It must also point out it's colorscheme .css

To daploy to the docker environment each app publish their latest
artefacts to the nginx html-root. Each app must make sure to put their
files into a folder by its own name. At the moment this is a bit
confusing. It's like the containers are not copying their files to the
static_files volume. Considering rewriting to a script that just does
docker cp to the volume.